refactor: is-production으로 이름 변경 (#447)
