+++ src/components/analytics/index.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import siteMetadata from 'data/siteMetadata.mjs';
 
-import { isProd } from '@/lib/isProduction';
+import { isProd } from '@/lib/is-production';
 
 import GA from './google-analytics';
